refactor: extract toppanel diagram overlays into seperate functionality
